+++ backend/src/main/java/cn/edu/zju/cs/jobmate/repositories/CompanyRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
  * DAO for {@link Company}.
  */
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer>, JpaSpecificationExecutor<Company> {
+public interface CompanyRepository extends JpaRepository<Company, Long>, JpaSpecificationExecutor<Company> {
 
     /**
      * Check if company exists by name.
